# Calculate the aligned bounding box for each particle; performance enhancements