… Legacy/bonej/src/main/java/org/bonej/plugins/ParticleDisplay.java