An axis-aligned bounding box has been implemented that takes the particle's principal axes as input and outputs the centroid and side length of the aligned bounding box. Users can choose to display numerical or 3D rendered results.

Addresses #288 

See the forum post for motivation and example output.
https://forum.image.sc/t/intersection-of-longest-principal-axis-and-object-surface/49105/8

Performance enhancements by
- multithreading analysis steps
- replacing Java Maps with Eclipse Collections primitive maps
- reducing neighbourhood handling during the first pass
- multithreading bucket fountain

Gains are most obvious on larger images (several GB) containing many particles; on very small images (2 MB) performance is similar or slower. Performance is also strongly affected by JRE optimisations, so 5 warm-up runs are recommended before taking measurements.

Using the s10up sample images:

pixels | master (ms) | aligned-box (ms) | nbh-decision-tree (ms) | multithread-bucket-fountain (ms)
-- | -- | -- | -- | --
2097152 | 52 | 59 | 55 | 82
2097152 | 62 | 58 | 60 | 55
2097152 | 45 | 52 | 70 | 48
2097152 | 70 | 93 | 48 | 46
2097152 | 54 | 59 | 50 | 53
16777216 | 216 | 140 | 141 | 172
16777216 | 171 | 193 | 163 | 157
16777216 | 146 | 165 | 183 | 189
16777216 | 196 | 137 | 128 | 158
16777216 | 202 | 186 | 211 | 144
134217728 | 1137 | 1002 | 1044 | 929
134217728 | 1070 | 1032 | 871 | 849
134217728 | 1134 | 1077 | 839 | 875
134217728 | 906 | 1290 | 1127 | 976
134217728 | 949 | 937 | 848 | 783
1073741824 | 8466 | 8014 | 7214 | 6318
1073741824 | 8250 | 6955 | 5989 | 6276
1073741824 | 8252 | 7133 | 6964 | 6239
1073741824 | 6621 | 6098 | 5652 | 6520
1073741824 | 6867 | 7361 | 5757 | 6426
7009835265 | 52688 | 50798 | 45447 | 48647
7009835265 | 50281 | 50944 | 41809 | 40914
7009835265 | 51344 | 45922 | 40795 | 43483
7009835265 | 54726 | 40896 | 37796 | 40491
7009835265 | 48201 | 38976 | 38965 | 35955
  |   **Average times (ms)**
2097152 | 56.6 | 64.2 | 56.6 | 56.8
16777216 | 186.2 | 164.2 | 165.2 | 164
134217728 | 1039.2 | 1067.6 | 945.8 | 882.4
1073741824 | 7691.2 | 7112.2 | 6315.2 | 6355.8
7009835265 | 51448 | 45507.2 | 40962.4 | 41898
  |   |  
**enhancements** | none | EC maps | EC maps, decision tree | all

Using the bglancy sample images


pixels | master (ms) | aligned-box (ms) | nbh-decision-tree (ms) | multithread-bucket-fountain (ms)
-- | -- | -- | -- | --
2097152 | 40 | 44 | 123 | 53
2097152 | 37 | 40 | 34 | 54
2097152 | 32 | 32 | 30 | 210
2097152 | 29 | 29 | 33 | 38
2097152 | 28 | 78 | 26 | 23
16777216 | 122 | 92 | 87 | 102
16777216 | 101 | 135 | 85 | 70
16777216 | 133 | 117 | 129 | 134
16777216 | 92 | 78 | 76 | 99
16777216 | 78 | 84 | 78 | 71
134217728 | 543 | 494 | 504 | 524
134217728 | 562 | 468 | 384 | 358
134217728 | 543 | 384 | 388 | 429
134217728 | 594 | 440 | 466 | 365
134217728 | 872 | 437 | 403 | 690
1073741824 | 5726 | 4562 | 4941 | 5877
1073741824 | 7204 | 3928 | 5313 | 4361
1073741824 | 6613 | 5002 | 5431 | 4917
1073741824 | 5372 | 3977 | 5199 | 5071
1073741824 | 5531 | 4213 | 4699 | 3536
8589934592 | 79427 | 49501 | 50670 | 41333
8589934592 | 69014 | 45375 | 46141 | 42721
8589934592 | 65565 | 49481 | 51481 | 45161
8589934592 | 56694 | 41562 | 41321 | 34487
8589934592 | 58830 | 41195 | 45756 | 37789
  | **Average Times (ms)**   |   |   |  
2097152 | 33.2 | 44.6 | 49.2 | 75.6
16777216 | 105.2 | 101.2 | 91 | 95.2
134217728 | 622.8 | 444.6 | 429 | 473.2
1073741824 | 6089.2 | 4336.4 | 5116.6 | 4752.4
7009835265 | 65906 | 45422.8 | 47073.8 | 40298.2
  |   |   |   |  
**enhancements** | none | EC maps | EC maps, decision tree | all